move highlighter to its own folder
