+++ scopes/react/component-highlighter/component-highlighter/bit-react-component.tsx
