+++ scopes/react/component-highlighter/component-highlighter/component-highlighter.docs.md
