+++ scopes/react/component-highlighter/component-highlighter/index.ts
@@ -1,0 +1,2 @@
+export { ComponentHighlighter } from './highlighter';
+export type { ComponentHighlightProps } from './highlighter';
